Add files via upload
JavaScript To-Do List App

Welcome to the JavaScript To-Do List App! This mini project is designed to showcase the concepts of functions, DOM handling, event handling, and DOM manipulation using HTML, CSS, and JavaScript.

Introduction
This project is a simple To-Do List application built with HTML, CSS, and JavaScript. It demonstrates the fundamental concepts of web development, including:

Function implementation
DOM (Document Object Model) handling
Event handling
DOM manipulation
Feel free to explore the code to understand how these concepts are applied to create a functional and interactive To-Do List.

Features
Add Tasks: Easily add tasks to your to-do list.
Edit Tasks: You can Edit a task
Delete Tasks: Remove tasks from the list.

If you'd like to contribute to this project, feel free to fork the repository and submit a pull request. Contributions, suggestions, and bug reports are welcome!